refactor: dedupe RPC banner header and settings-tab test scaffolding
/simplify 4-agent 리뷰에서 두 관점(simplification, altitude)이 독립적으로
같은 지점을 지적해 정리한다.

- renderRpcSetupBannerHeader() 추출: 배너 제목 div(role="heading")와
  설명 문단이 connection-card / credential-form 두 렌더러에 verbatim
  중복돼 있었다. ARIA 절충의 근거 주석도 한 곳에서만 관리한다.
- createSettingsTabHarness() 추출: 두 settings-tab 스펙이 동일한
  5-인자 생성자 호출을 각각 들고 있었다. tab은 untyped로 반환해
  각 스펙이 자신이 검증하는 private 표면으로만 캐스팅한다.

동작 변경 없음 — 순수 이동. build/lint/800 tests 통과.
